% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.file}
\alias{extract.file}
\title{Extract participant ID from TCGA barcode}
\usage{
extract.file(tar.file, extract.file, new.file, resultsdir)
}
\value{
file details
}
\description{
Extract participant ID from TCGA barcode
}
