% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.participant}
\alias{extract.participant}
\title{Extract participants ID from TGCA barcode}
\usage{
extract.participant(id)
}
\arguments{
\item{id}{TGCA barcode}
}
\value{
A sstring representing participant ID
}
\description{
Extracts the participant identifier from a TCGA sample barcode
}
