% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{my.write.table}
\alias{my.write.table}
\title{Write a table}
\usage{
my.write.table(x, filename)
}
\arguments{
\item{x}{dataframe to save in a table}

\item{filename}{filename for table}
}
\value{
none
}
\description{
Writes a dataframe to a table
}
