% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{my_write_table}
\alias{my_write_table}
\title{Write table}
\usage{
my_write_table(x, filename)
}
\arguments{
\item{x}{data framte to write}

\item{filename}{output file name}
}
\value{
none
}
\description{
?
}
