% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.file}
\alias{extract.file}
\title{File extraction
Extracts the TCGA tar.gz to a named output}
\usage{
extract.file(tar.file, extract.file, new.file, resultsdir)
}
\arguments{
\item{tar.file}{tar.gz file containing TCGA}

\item{extract.file}{file which enables the extraction}

\item{new.file}{name of desired output file}

\item{resultsdir}{location of desired output file}
}
\value{
A new extracted file in the specified location
}
\description{
File extraction
Extracts the TCGA tar.gz to a named output
}
