% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.file}
\alias{extract.file}
\title{Extract files from the .tar archive and move them to the results directory}
\usage{
extract.file(tar.file, extract.file, new.file, resultsdir)
}
\arguments{
\item{tar.file}{.tar archive containing files of interest}

\item{extract.file}{files from the .tar archive to extract}

\item{new.file}{file to move extracted data files into}

\item{resultsdir}{where the new.file will be saved}
}
\value{
untarred data file moved to resultsdir/new.file
}
\description{
Extract files from the .tar archive and move them to the results directory
}
