% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.participant}
\alias{extract.participant}
\title{Participant extraction
This function will extract participants}
\usage{
extract.participant(id)
}
\arguments{
\item{id}{TCGA barcode}
}
\value{
participant ID
}
\description{
Participant extraction
This function will extract participants
}
