% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{my.write.table}
\alias{my.write.table}
\title{Write a table detailing the required outputs}
\usage{
my.write.table(x, filename)
}
\arguments{
\item{x}{}

\item{filename}{}
}
\value{
table
}
\description{
Write a table detailing the required outputs
}
