% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{my.write.table}
\alias{my.write.table}
\title{Create a table and save it to a directory}
\usage{
my.write.table(x, filename)
}
\arguments{
\item{x}{dataframe to save}

\item{filename}{path and filename to save the dataframe to}
}
\value{
a saved dataframe
}
\description{
Create a table and save it to a directory
}
